iRedMail/tools/scripts/requirements.txt
#
####### requirements.txt #######
#
####### pip install -r requirements.txt #######
#
###### Requirements without Version Specifiers ######
ArgumentParser
mysql-connector
#
